week 6/requirements.txt
pandas==3.0.2
billboard.py==7.1.0
